ubuntu: install rcconf
rcconf is Debian runlevel configuration tool. Rcconf allows you to control
which services are started when the system boots up or reboots. It displays
a menu of all the services which could be started at boot. The ones that
are configured to do so are marked and you can toggle individual services
on and off.

Signed-off-by: Fudong Bai <fudongbai@gmail.com>